flytt byggFulltNavn og bruk i tittel
