+++ app/components/personlinje/Personlinje.helpers.ts
@@ -1,19 +1,6 @@
 import { differenceInYears } from "date-fns";
 
 /**
- * Bygger fullt navn fra fornavn, mellomnavn og etternavn
- * Håndterer tilfeller hvor mellomnavn kan være null/undefined
- */
-export function byggFulltNavn(
-  fornavn: string,
-  mellomnavn: string | null | undefined,
-  etternavn: string,
-): string {
-  const navneDeler = [fornavn, mellomnavn, etternavn].filter(Boolean);
-  return navneDeler.join(" ");
-}
-
-/**
  * Beregner alder basert på fødselsdato
  * Returnerer antall hele år mellom fødselsdato og i dag
  */
